Temp commit - Part 1 of ?? Refactoring.
