+++ src/com/tracv/model/State.java
@@ -1,0 +1,5 @@
+package com.tracv.model;
+
+public enum State {
+    PLAYING, TERMINATED, PAUSED
+}
