+++ src/com/tracv/ui/TDFrame.java
@@ -1,16 +1,16 @@
 package com.tracv.ui;
 
+import com.tracv.swing.Frame;
 import com.tracv.util.*;
 
 import javax.swing.*;
-import java.awt.*;
 import java.awt.event.WindowEvent;
 import java.awt.event.WindowFocusListener;
 
 /**
  * The Window containing all other components to the program
  */
-public class TDFrame extends JFrame implements WindowFocusListener {
+public class TDFrame extends Frame implements WindowFocusListener {
 
     //The various content panels.
     private MainPane mainPane;
@@ -21,48 +21,34 @@ public class TDFrame extends JFrame implements WindowFocusListener {
     //GlassPane
     private MenuPane menuPane;
 
+    //Holds previous pane allows us to jump back
     private JComponent last;
 
+    //Keeps mouse on the map.
     private MouseHooker mouseHooker;
 
+    /**
+     * Initializes program.
+     */
     public TDFrame() {
+        super();
         mouseHooker = new MouseHooker(this);
+        this.addWindowFocusListener(this);
 
         //Create Content Panes.
         mainPane = new MainPane(this);
         tdGame = new TDGame(this);
-        levelSelectPane = new LevelSelectPane(this);
-        menuPane = new MenuPane(this);
+        levelSelectPane = new LevelSelectPane(this, tdGame);
+        menuPane = new MenuPane(this, tdGame);
         settingsPane = new SettingsPane(this);
-
-        this.setUndecorated(true);
-        //this.setBackground(new Color(1.0f, 1.0f, 1.0f, 0.8f));
 
         this.setContentPane(mainPane);
         this.setGlassPane(menuPane);
-
-        this.setSize(Constants.FRAME_DEFAULT_SIZE);
-        Comp.center(this);
-
-        this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.addWindowFocusListener(this);
-        this.setVisible(true);
     }
 
     public static void main(String args[]) {
         new TDFrame(); //Starts the program by calling TDFrame's constructor
         Logger.getInstance().log("Program Startup Complete", LoggerLevel.STATUS);
-    }
-
-    public void restartGame() {
-        tdGame.restart();
-    }
-
-    public void newGame(int i){
-        tdGame.startNewGame(i);
-        if(this.getContentPane() != tdGame) {
-            switchToGamePanel();
-        }
     }
 
     public void switchToMainPanel() {
@@ -75,6 +61,7 @@ public class TDFrame extends JFrame implements WindowFocusListener {
         switchPanel(tdGame);
         tdGame.grabFocus(); // Give panel focus.
     }
+
     public void switchToSettingsPanel(){
         switchPanel(settingsPane);
     }
@@ -97,24 +84,24 @@ public class TDFrame extends JFrame implements WindowFocusListener {
      * @param msg
      */
     public void toggleMenu(boolean b, String msg) {
-        System.out.println("Menu " + b);
         if(b){
-            menuPane.showMenu((JComponent) this.getContentPane(), msg);
+            menuPane.showMenu((JComponent)this.getContentPane(), msg);
             mouseHooker.setActive(false);
         }else {
             menuPane.setVisible(false);
 
-            //If resuming game, resume game.
+            //If switching back to game, must grab mouse binder.
             if(this.getContentPane() == tdGame){
-                System.out.println("Rebinding hooker!!");
-                tdGame.resumeGame();
                 mouseHooker.setActive(true);
             }
-
         }
         SwingUtilities.invokeLater(()-> this.validate());
     }
 
+    /**
+     * Default toggle menu.
+     *
+     */
     public void toggleMenu(boolean b) {
         toggleMenu(b, null);
     }
@@ -124,6 +111,7 @@ public class TDFrame extends JFrame implements WindowFocusListener {
     }
 
 
+    //Adjust MouseHooker based off if we have focus on the window
     @Override
     public void windowGainedFocus(WindowEvent e) {
         if(TDFrame.this.getContentPane() == tdGame){
